Virtuoso specific geo-literals for points now generated with float precision rather than double as virtuoso converts to float implicitely.
git-svn-id: http://linkedgeodata.googlecode.com/svn/trunk@228 00633da9-98c1-349d-95eb-89b68c819386

+++ src/main/java/org/linkedgeodata/osm/mapping/impl/SimpleNodeToRDFTransformer.java
@@ -113,7 +113,7 @@ public class SimpleNodeToRDFTransformer
 	
 	public static Literal generateVirtuosoLiteral(Point2D point)
 	{
-		Literal result = ResourceFactory.createTypedLiteral("POINT(" + point.getX() + " " + point.getY() + ")", virtrdfGeometry);
+		Literal result = ResourceFactory.createTypedLiteral("POINT(" + (float)point.getX() + " " + (float)point.getY() + ")", virtrdfGeometry);
 		return result;
 	}
